+++ Application/LCD/LCD_BasicGraphics.h
@@ -265,7 +265,7 @@ void LCD_BoldRoundFrameTransp		(uint32_t posBuff, uint32_t BkpSizeX,uint32_t Bkp
 void LCD_LittleRoundRectangle(uint32_t posBuff, uint32_t BkpSizeX,uint32_t BkpSizeY, uint32_t x,uint32_t y, uint32_t width, uint32_t height, uint32_t FrameColor, uint32_t FillColor, uint32_t BkpColor);
 void LCD_LittleRoundFrame(uint32_t posBuff, uint32_t BkpSizeX,uint32_t BkpSizeY, uint32_t x,uint32_t y, uint32_t width, uint32_t height, uint32_t FrameColor, uint32_t FillColor, uint32_t BkpColor);
 
-structPosition DrawLine(uint32_t posBuff,uint16_t x, uint16_t y, uint16_t len, uint16_t degree, uint32_t lineColor,uint32_t BkpSizeX, float ratioAA1, float ratioAA2 ,uint32_t bk1Color, uint32_t bk2Color);
+structPosition DrawLine(uint32_t posBuff,uint16_t x, uint16_t y, uint16_t len, float degree, uint32_t lineColor,uint32_t BkpSizeX, float ratioAA1, float ratioAA2 ,uint32_t bk1Color, uint32_t bk2Color);
 
 void LCD_SetCircleParam(float outRatio, float inRatio, int len, ...);
 void LCD_SetCircleDegrees(int len, ...);
